Changed list design in helpfoss.md
Changed a couple things at the top. I thought it'd make more sense if the list header had a more specific name so people who don't know what FOSS is can easily understand. Also, I formatted the list to start naming the items from top to bottom in category instead of alphabetically by software name. I felt it was easier to navigate from top to bottom and to know from the get-go specifically what software is for what instead of having to needlessly scroll down to investigate. I added the software into a blockquote list series because once you add more software to the original list, in the long run it will just take up a lot more unnecessary space. Hope this helps to be a little more user friendly :).